New Timer Functionality and Page Themes
The timer now only takes input for Total Seconds and then converts to minute:second format.  When time runs out, it goes to next question and resets.  Added dark green and gold theme to all pages, and changed the fonts to be consistent.  Font doesn't have to be what I made it, but at least now it is consistent throughout.